14 MenuController & menu.html

+++ src/main/kotlin/com/aslanjavasky/shawarmadelviry/domain/model/MenuItem.kt
@@ -1,8 +1,10 @@
 package com.aslanjavasky.shawarmadelviry.domain.model
 
+import java.math.BigDecimal
+
 data class MenuItem(
     var id: Long,
     var name: String,
     var menuSection: MenuSection,
-    var price: Int
+    var price: BigDecimal
 )
